migrated to NBench 1.2.1 to help resolve #4
