+++ .gitignore
@@ -46,6 +46,7 @@ _site/
 
 # MSTest test Results
 [Tt]est[Rr]esult*/
+[Pp]erf[Rr]esult*/
 [Bb]uild[Ll]og.*
 
 # NUNIT
