+++ src/Petabridge.Tracing.Zipkin.Tests.Performance/Petabridge.Tracing.Zipkin.Tests.Performance.csproj
@@ -3,12 +3,13 @@
 
 
   <PropertyGroup>    
-    <TargetFramework>net461</TargetFramework>
+    <TargetFrameworks>net461;netcoreapp2.0</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(TestSdkVersion)" />
-    <PackageReference Include="NBench" Version="1.0.4" />
+    <PackageReference Include="NBench" Version="1.2.1" />
+    <DotNetCliToolReference Include="dotnet-nbench" Version="1.2.1" />
   </ItemGroup>
 
   <ItemGroup>
